--- Convert-Word-to-image/.NET-Standard/Convert-Word-to-image/Convert-Word-to-image.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Convert_Word_to_image</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
-  </ItemGroup>
-
-</Project>
