+++ First-page-of-Word-to-image/.NET/First-page-of-Word-to-image/First-page-of-Word-to-image.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <RootNamespace>First_page_of_Word_to_image</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
+  </ItemGroup>
+
+</Project>
