+++ Specific-range-of-pages-to-image/.NET-Framework/Specific-range-of-pages-to-image/packages.config
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Syncfusion.Compression.Base" version="19.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.DocIO.WinForms" version="19.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.Licensing" version="19.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.OfficeChart.Base" version="19.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.OfficeChartToImageConverter.WinForms" version="19.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.SfChart.WPF" version="19.4.0.50" targetFramework="net46" />
+  <package id="Syncfusion.Compression.Base" version="26.1.42" targetFramework="net462" />
+  <package id="Syncfusion.DocIO.WinForms" version="26.1.42" targetFramework="net462" />
+  <package id="Syncfusion.Licensing" version="26.1.42" targetFramework="net462" />
+  <package id="Syncfusion.OfficeChart.Base" version="26.1.42" targetFramework="net462" />
+  <package id="Syncfusion.OfficeChartToImageConverter.WinForms" version="26.1.42" targetFramework="net462" />
+  <package id="Syncfusion.SfChart.WPF" version="26.1.42" targetFramework="net462" />
 </packages>
