--- Specific-range-of-pages-to-image/.NET-Standard/Specific-range-of-pages-to-image/Specific-range-of-pages-to-image.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Specific_range_of_pages_to_image</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
-  </ItemGroup>
-
-</Project>
