+++ Specific-range-of-pages-to-image/.NET/Specific-range-of-pages-to-image/Specific-range-of-pages-to-image.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <RootNamespace>Specific_range_of_pages_to_image</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Syncfusion.DocIORenderer.Net.Core" Version="*" />
+  </ItemGroup>
+
+</Project>
